Use unmodified PDF if OCR fails

+++ watch.sh
@@ -16,19 +16,24 @@ do
 
             ocrmypdf -l deu+eng --sidecar "/tmp/ocr/${filename}.txt" "${pathAndFilename}" "/tmp/ocr/${filename}.pdf"
 
-            documentNumber=$(head -n 1 "/tmp/ocr/${filename}.txt" | grep -E '^\s*[0-9]{7}\s*$')
-            if [[ -z "${documentNumber}" ]]; then
-                targetFilename=${filename}
+            if [[ -f "/tmp/ocr/${filename}.txt" ]]; then
+                documentNumber=$(head -n 1 "/tmp/ocr/${filename}.txt" | grep -E '^\s*[0-9]{7}\s*$')
+                if [[ -z "${documentNumber}" ]]; then
+                    targetFilename=${filename}
+                else
+                    targetFilename=${filenamePrefix}-${documentNumber}.pdf
+                fi
+
+                chmod 666 "/tmp/ocr/${filename}.pdf"
+                mv -f "/tmp/ocr/${filename}.pdf" "/data/output/${targetFilename}"
+                rm "${pathAndFilename}"
+                rm "/tmp/ocr/${filename}.txt"
+
+                echo "Finished ocr, result is ${targetFilename}"
             else
-                targetFilename=${filenamePrefix}-${documentNumber}.pdf
+                echo "Text recognition failed, copying file without OCR to ${targetFilename}"
+                mv -f "${pathAndFilename}.pdf" "/data/output/${targetFilename}"
             fi
-
-            chmod 666 "/tmp/ocr/${filename}.pdf"
-            mv -f "/tmp/ocr/${filename}.pdf" "/data/output/${targetFilename}"
-            rm "${pathAndFilename}"
-            rm "/tmp/ocr/${filename}.txt"
-
-            echo "Finished ocr, result is ${targetFilename}"
         done
     sleep 5
     echo -ne "."
